Import intro materials from the wiki
Imported several pages from the splunk-sdk-python.wiki hosted on GitHub, this
provides additional context that's quite informative and hopefully becomes more
visible by being included in the docs rather than the wiki.  (I only recently
discovered that the wiki was enabled.)   Not all the wiki pages were imported,
just the inactive pages.  The last modification time on all the pages imported
were from 2014 or earlier.

 * Architecture
 * Branching Model
 * FAQ
 * Contact Us

The only page not imported was the Release doc, which seems to be more active
on the wiki, and seems less relevant to average users.

If this is accepted in the python docs, then it make sense that the wiki pages
should be removed and/or updated to point to the read-the-docs version instead.

From browsing the commit logs, looks like most of the credit for this content
goes to Shakeel Mohamed.

Docs were converted from Markdown to ReST using pandoc.  Assuming others like
this idea, inter doc links (to py objects) could be improved.